TaskFlow

+++ src/TaskFlow.Data/Repositories/Tasks/ITaskRepository.cs
@@ -6,8 +6,8 @@ namespace TaskFlow.Data.Repositories.Tasks
     {
         void Add(MyTask task);
         List<MyTask> GetAll();
-        MyTask Get(int id);
+        MyTask Get(long id);
         void Update(MyTask task);
-        bool Delete(int id);
+        bool Delete(long id);
     }
 }
